Initialize repo for rtc-video-quality.
We will be using google/rtc-video-quality for metric code analysis.
This tool will be used to measure the quality metrics for the different
video encoders that will be used in this project.

Since we will be making changes to this tool, we will be copying
the source code of rtc-video-quality into this project and make changes
to it in this repo. This commit is just to initialize the base repo so
that the we can keep track of new changes that will be made.