Convert python2 code to python3.
Since python2 is deprecated and no longer supported, we will be
converting the all the code used to python3. Used a tool (2to3) to
convert most of the print/dictionary statements to python3. With
python3,some of the functions returned bytes by default instead of
string. Thus we had to manually add encoding to utf-8 in subprocess commands.