Add offline configurations for SVT-AV1. (#11)
* Add offline configurations for SVT-AV1.

In order to run SVT 2 pass, you need to run 2 seperate commands and use
a statfile. Thus, the command for Popen was changed to use shell so that
it supports shell symbols such as '&&'. So currently, svt_command can
return a command of the form command1 + ['&&'] + command2 where command1
and command2 are list of strings.

* use shlex to run string commands

* check for equality instead of substring replacement in str comp.